refactor: change svg tags file name
