+++ src/lib/svg_tags.js
